Make data writer interface to send received data to. Make kinesis queue implementation of this interface.
